fix(#2456): split impact_model.py (1133 LOC) into three focused modules
- _impact_physics.py (664 LOC): enums, dataclasses, ImpactModel ABC,
  RigidBodyImpactModel, SpringDamperImpactModel, FiniteTimeImpactModel,
  compute_gear_effect_spin, validate_energy_balance, create_impact_model
- _impact_recorder.py (495 LOC): ImpactEvent, ImpactRecorder, ImpactSolverAPI
- impact_model.py (55 LOC): thin coordinator facade re-exporting all public names

All 13 public names remain importable from impact_model (backward compat).
Adds TDD contract tests: structure, LOC budget, and public API import checks.
